change the data channel to RTM.

+++ edusdk-impl/src/main/kotlin/io/agora/rte/IRteChannel.kt
@@ -5,7 +5,7 @@ import io.agora.rtc.models.ChannelMediaOptions
 import io.agora.rtm.ResultCallback
 
 interface IRteChannel {
-    fun join(token: String, rtmToken: String, rtcUid: Int, rtmUid: String,
+    fun join(token: String, rtmToken: String, rtcUid: Long, rtmUid: String,
              mediaOptions: ChannelMediaOptions, @NonNull callback: ResultCallback<Void>)
 
     fun leave()
